backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
email-validator>=2.1.0

# Database
SQLAlchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg>=0.29.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-jose>=3.3.0

# XRPL integration
xrpl-py>=2.4.0
websockets>=12.0

# AWS
boto3>=1.29.3
botocore>=1.32.3

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.1
pytest-cov>=4.1.0

# Logging
loguru>=0.7.2

# Utilities
pydantic==2.5.1
pydantic-settings>=2.1.0
python-dotenv==1.0.0
tenacity>=8.2.3
aiohttp>=3.9.0
requests>=2.31.0

# Development tools
black>=23.11.0
flake8>=6.1.0
isort>=5.12.0
mypy>=1.7.0 